+++ node/src/types/kernel/api/export/exportResources/response.d.ts
@@ -15,6 +15,10 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+/**
+ * Packages the specified files and folders as *.zip files and returns their download
+ * addresses
+ */
 export interface IResponse {
     /**
      * status code
@@ -32,7 +36,7 @@ export interface IResponse {
  */
 export interface IData {
     /**
-     * the path of output file under the workspace
+     * the path of *.zip exported under the workspace
      */
     path: string;
 }